MsgTree: fix issue of tree in MODE_DEFER being modified after walk()
MODE_DEFER is the default mode of MsgTree and allow "preloading" the
tree without immediately having to assign keys (nodes) to last lines.
Keys assignment is made when walk() is first called. After this, a mode
change to MODE_SHIFT was missing, leading to duplicate keys being
present in the tree when messages were added afterwards. This behavior
is present in Gateway.py which is using MsgTree to perform its
"grooming" message aggregation.

Change-Id: I49cef08d1b70cf06b28e8894f0e525250f885a7c