Worker: add new StreamWorker class
The StreamWorker manages a single group of streams (using one
EngineClient internally). These streams can be configured in read or
write modes. StreamWorker doesn't execute any commands, the streams have
to be already established.

StreamWorker is a now an important class for Tree mode support as it is
used directly in the Gateway.py module. The gateway acts as a
StreamWorker reading from stdin (from the parent ssh process) and
writing to stdout (back to the parent ssh process).

It is now also used as a base class for WorkerSimple/WorkerPopen to
factorize code.

This commit also renames previously introduced Task default option
'worker' to 'distant_worker'.

Change-Id: Icfd6613e06e01cdc250510560acbbc39c5554035