MsgTree: use bytes buffers instead of string (#337)
To support Python 3, we are moving towards using bytes everywhere in
ClusterShell instead of strings. Encoding is always attached to
Python 3's strings, and we don't want to deal with that. The 'bytes'
type is very similar to Python 2's str. Popen returns bytes in Python 3
for example.  We need to change MsgTree so that lines are in bytes.
For Python 2, this doesn't change anything as bytes are str.

- use bytes (buffer) instead of str
- keep the MsgTreeElem.message() method that now works with both Python
- deprecate the use of MsgTreeElem.__str__() in Python 2, as it cannot be
  used anymore in Python 3
- add MsgTreeElem.__bytes__() method (for bytes() casting) in Python 3
- change MsgTree tests to use bytes

Change-Id: I5b7917c6e2c025e6eaf3afe4c3776f7a0beb70b3