Avoid leaking open file descriptors during tests
nosetests seems to keep a reference on all unittest.TestCase instances,
so it is important to take that into account to avoid leaking file
descriptors during a test suite run.

This change fixes the tests and fastsubprocess after thoroughly
analyzing the situation of leaking file descriptors.

The test suite doesn't leak file descriptors anymore when this patch is
applied.

Change-Id: I0983e3684ff13f0dd277bd14787ff416b8c9365b