NodeSet: implement node wildcards (#349)
Node wildcard expansion, using the well-known characters '*' and '?',
allows users to filter on node names that match the provided wildcard
mask.

The implementation modifies ParsingEngine to search for nodes from 'all'
using the provided wildcard mask, that is, nodes that are returned by
the all upcall, or in case of absence, the union of list+map.

Simultaneous use of wildcards and ranges (like '*foo[1-3]') is fully
supported.

With this change,
    $ nodeset -f '*'
becomes the same as:
    $ nodeset -f -a

Closes #349

Change-Id: I1b99811dc65bd7fd9a5a9fedc6e5f1f3b8cb875e
Signed-off-by: Stephane Thiell <sthiell@stanford.edu>
Signed-off-by: Aurelien Degremont <aurelien.degremont@cea.fr>