(#370) NodeSet: speed-up nodeset parsing
Rewrite NodeSet._next_op() using re.search(), testing
all supported operators with only one call instead of calling
str.find() for each operator (we support 4 of them)
which is more expensive.

ParsingEngine.OP_CODES dict has been swapped as we now need
an index using a operator character as key.

This is useful when pattern to parse has a lot of operators
(ie: foo1,foo2,foo3,...,foo5000)

Benchmark shows a 60% improvement with this patch when parsing
large strings with a lot of operators.

Change-Id: Idca7763afb34269210799e8a5176785c314382da