NodeSet: add special notation @@source to expand group names
This change allows access to all group names of a specific source in
node set expressions. The special operator `@@` optionally followed
by a source expands group names that can then be manipulated as a set.
The default group source is used when `@@` is used alone.

Note: In that case, group names are not prefixed with @ and as such, not
further expanded into nodes by NodeSet.